feat(tasks): Fetch tasks from backend in `task-list` component.
- Move NumberInput component to separate file
- Add dividers between tasks
